Move copied code out
