+++ src/execute.rs
@@ -1,20 +1,17 @@
+use crate::ethereum::{EthEvmExecutor, EthExecuteOutput};
 use crate::gnosis::{apply_block_rewards_contract_call, apply_withdrawals_contract_call};
 use reth::{
     api::ConfigureEvm,
     primitives::{
         Address, BlockNumber, BlockWithSenders, ChainSpec, Header, PruneModes, Receipt, Receipts,
-        Request, U256,
+        U256,
     },
     providers::ProviderError,
     revm::{
         batch::{BlockBatchRecord, BlockExecutorStats},
         db::states::bundle_state::BundleRetention,
-        primitives::{BlockEnv, CfgEnvWithHandlerCfg, EnvWithHandlerCfg, ResultAndState},
-        state_change::{
-            apply_beacon_root_contract_call, apply_blockhashes_update,
-            apply_withdrawal_requests_contract_call,
-        },
-        Database, DatabaseCommit, Evm, State,
+        primitives::{BlockEnv, CfgEnvWithHandlerCfg, EnvWithHandlerCfg},
+        Database, State,
     },
 };
 use reth_ethereum_consensus::validate_block_post_execution;
@@ -22,119 +19,7 @@ use reth_evm::execute::{
     BatchBlockExecutionOutput, BatchExecutor, BlockExecutionError, BlockExecutionInput,
     BlockExecutionOutput, BlockExecutorProvider, BlockValidationError, Executor,
 };
-use reth_evm_ethereum::eip6110::parse_deposits_from_receipts;
 use std::{collections::HashMap, sync::Arc};
-
-/// Helper container type for EVM with chain spec.
-// [Gnosis/fork] Copy paste code from crates/ethereum/evm/src/execute.rs::EthBatchExecutor
-#[derive(Debug, Clone)]
-struct GnosisEvmExecutor<EvmConfig> {
-    /// The chainspec
-    chain_spec: Arc<ChainSpec>,
-    /// How to create an EVM.
-    evm_config: EvmConfig,
-}
-
-// [Gnosis/fork] Copy paste code from crates/ethereum/evm/src/execute.rs::EthBatchExecutor
-impl<EvmConfig> GnosisEvmExecutor<EvmConfig>
-where
-    EvmConfig: ConfigureEvm,
-{
-    // [Gnosis/fork] Copy paste code from crates/ethereum/evm/src/execute.rs::EthEvmExecutor
-    fn execute_state_transitions<Ext, DB>(
-        &self,
-        block: &BlockWithSenders,
-        mut evm: Evm<'_, Ext, &mut State<DB>>,
-    ) -> Result<EthExecuteOutput, BlockExecutionError>
-    where
-        DB: Database<Error = ProviderError>,
-    {
-        // apply pre execution changes
-        apply_beacon_root_contract_call(
-            &self.chain_spec,
-            block.timestamp,
-            block.number,
-            block.parent_beacon_block_root,
-            &mut evm,
-        )?;
-        apply_blockhashes_update(
-            evm.db_mut(),
-            &self.chain_spec,
-            block.timestamp,
-            block.number,
-            block.parent_hash,
-        )?;
-
-        // execute transactions
-        let mut cumulative_gas_used = 0;
-        let mut receipts = Vec::with_capacity(block.body.len());
-        for (sender, transaction) in block.transactions_with_sender() {
-            // The sum of the transaction‚Äôs gas limit, Tg, and the gas utilized in this block prior,
-            // must be no greater than the block‚Äôs gasLimit.
-            let block_available_gas = block.header.gas_limit - cumulative_gas_used;
-            if transaction.gas_limit() > block_available_gas {
-                return Err(
-                    BlockValidationError::TransactionGasLimitMoreThanAvailableBlockGas {
-                        transaction_gas_limit: transaction.gas_limit(),
-                        block_available_gas,
-                    }
-                    .into(),
-                );
-            }
-
-            EvmConfig::fill_tx_env(evm.tx_mut(), transaction, *sender);
-
-            // Execute transaction.
-            let ResultAndState { result, state } = evm.transact().map_err(move |err| {
-                // Ensure hash is calculated for error log, if not already done
-                BlockValidationError::EVM {
-                    hash: transaction.recalculate_hash(),
-                    error: err.into(),
-                }
-            })?;
-            evm.db_mut().commit(state);
-
-            // append gas used
-            cumulative_gas_used += result.gas_used();
-
-            // Push transaction changeset and calculate header bloom filter for receipt.
-            receipts.push(
-                #[allow(clippy::needless_update)] // side-effect of optimism fields
-                Receipt {
-                    tx_type: transaction.tx_type(),
-                    // Success flag was added in `EIP-658: Embedding transaction status code in
-                    // receipts`.
-                    success: result.is_success(),
-                    cumulative_gas_used,
-                    // convert to reth log
-                    logs: result.into_logs(),
-                    ..Default::default()
-                },
-            );
-        }
-
-        let requests = if self
-            .chain_spec
-            .is_prague_active_at_timestamp(block.timestamp)
-        {
-            // Collect all EIP-6110 deposits
-            let deposit_requests = parse_deposits_from_receipts(&self.chain_spec, &receipts)?;
-
-            // Collect all EIP-7685 requests
-            let withdrawal_requests = apply_withdrawal_requests_contract_call(&mut evm)?;
-
-            [deposit_requests, withdrawal_requests].concat()
-        } else {
-            vec![]
-        };
-
-        Ok(EthExecuteOutput {
-            receipts,
-            requests,
-            gas_used: cumulative_gas_used,
-        })
-    }
-}
 
 #[derive(Debug, Clone)]
 pub struct GnosisExecutorProvider<EvmConfig: Clone> {
@@ -207,7 +92,7 @@ where
 #[derive(Debug)]
 pub struct GnosisBlockExecutor<EvmConfig, DB> {
     /// Chain specific evm config that's used to execute a block.
-    executor: GnosisEvmExecutor<EvmConfig>,
+    executor: EthEvmExecutor<EvmConfig>,
     /// The state to use for execution
     state: State<DB>,
 }
@@ -217,7 +102,7 @@ impl<EvmConfig, DB> GnosisBlockExecutor<EvmConfig, DB> {
     /// Creates a new Ethereum block executor.
     pub fn new(chain_spec: Arc<ChainSpec>, evm_config: EvmConfig, state: State<DB>) -> Self {
         Self {
-            executor: GnosisEvmExecutor {
+            executor: EthEvmExecutor {
                 chain_spec,
                 evm_config,
             },
@@ -351,15 +236,6 @@ where
 
         Ok(())
     }
-}
-
-/// Helper type for the output of executing a block.
-// [Gnosis/fork] Copy paste code from crates/ethereum/evm/src/execute.rs::EthExecuteOutput
-#[derive(Debug, Clone)]
-struct EthExecuteOutput {
-    receipts: Vec<Receipt>,
-    requests: Vec<Request>,
-    gas_used: u64,
 }
 
 // Trait required by BlockExecutorProvider associated type Executor
